reset DB and change fields to decimal

+++ Marinel/Data/Entities/Expense.cs
@@ -9,7 +9,7 @@ namespace SchoolDraftWebsite.Data.Entities
         public DateTime Date { get; set; }
         public string Description { get; set; }
         public int Quantity { get; set; }
-        public long Amount { get; set; }
-        public long Total { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Total { get; set; }
     }
 }
